yetki kontrolleri
yetki kontrolleri

+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/controller/HomeController.java
@@ -2,6 +2,7 @@ package com.udacity.jwdnd.course1.cloudstorage.controller;
 
 
 import com.udacity.jwdnd.course1.cloudstorage.config.StoragePropertiesConfig;
+import com.udacity.jwdnd.course1.cloudstorage.exception.AuthorizationException;
 import com.udacity.jwdnd.course1.cloudstorage.exception.StorageException;
 import com.udacity.jwdnd.course1.cloudstorage.mapper.CredentialMapper;
 import com.udacity.jwdnd.course1.cloudstorage.mapper.FileMapper;
@@ -76,6 +77,7 @@ public class HomeController {
         model.addAttribute("login_user_name", username);
 
         var messageMap = new HashMap<String, String>();
+        messageMap.put("authException", "invalid authorization");
         messageMap.put("aCred", "credential added");
         messageMap.put("uCred", "credential updated");
         messageMap.put("dCred", "credential deleted");
@@ -98,28 +100,42 @@ public class HomeController {
         int userId = user.getUserId();
         String title = noteTitle;
         String description = noteDescription;
-        int noteSize = noteMapper.insert(new Note(title, description, userId));
+        noteMapper.insert(new Note(title, description, userId));
         return "redirect:/home?msg=aNote";
     }
 
     @GetMapping("/delete-note")
-    public String deleteNote(@RequestParam("noteId") String noteId) {
+    public String deleteNote(@RequestParam("noteId") String noteId, Authentication authentication) {
         int noteid = Integer.parseInt(noteId);
+        Note note = noteMapper.getNote(noteid);
+        if (!checkAuthorization(note.getUserId(), authentication)) {
+            return "redirect:/home?msg=authException";
+        }
         noteMapper.delete(noteid);
         return "redirect:/home?msg=dNote";
     }
 
     @GetMapping("/edit-note")
-    public String editNote(@RequestParam("noteId") String noteId, Model model) {
+    public String editNote(@RequestParam("noteId") String noteId, Model model, Authentication authentication) {
         int noteid = Integer.parseInt(noteId);
         Note note = noteMapper.getNote(noteid);
+
+        if (!checkAuthorization(note.getUserId(), authentication)) {
+            return "redirect:/home?msg=authException";
+        }
+
         model.addAttribute("note", note);
         return "edit-note";
     }
 
     @PostMapping("/update-note")
-    public String updateNote(@ModelAttribute("note") Note note) {
+    public String updateNote(@ModelAttribute("note") Note note, Authentication authentication) {
         noteMapper.update(note);
+
+        if (!checkAuthorization(note.getUserId(), authentication)) {
+            return "redirect:/home?msg=authException";
+        }
+
         return "redirect:/home?msg=uNote";
     }
 
@@ -128,45 +144,51 @@ public class HomeController {
                                 @RequestParam String username,
                                 @RequestParam String password,
                                 Authentication authentication, Model model) {
-        User user = userService.getUser(authentication.getName());
-        int userId = user.getUserId();
         SecureRandom random = new SecureRandom();
         byte[] key = new byte[16];
         random.nextBytes(key);
         String encodedKey = Base64.getEncoder().encodeToString(key);
         String encryptedPassword = encryptionService.encryptValue(password, encodedKey);
-        int credentialSize = credentialMapper.insert(new Credential(url, username, encodedKey, encryptedPassword, userId));
+        int credentialSize = credentialMapper.insert(new Credential(url, username, encodedKey, encryptedPassword, userService.getUserId(authentication.getName())));
         return "redirect:/home?msg=aCred";
     }
 
     @GetMapping("/delete-credential")
-    public String deleteCredential(@RequestParam("credentialId") String credentialId) {
+    public String deleteCredential(@RequestParam("credentialId") String credentialId, Authentication authentication) {
         int credentialid = Integer.parseInt(credentialId);
+        Credential credential = credentialMapper.getCredential(credentialid);
+        if (!checkAuthorization(credential.getUserId(), authentication)) {
+            return "redirect:/home?msg=authException";
+        }
         credentialMapper.delete(credentialid);
         return "redirect:/home?msg=dCred";
     }
 
     @GetMapping("/edit-credential")
-    public String editCredential(@RequestParam("credentialId") String credentialId, Model model) {
+    public String editCredential(@RequestParam("credentialId") String credentialId, Model model, Authentication authentication) {
         int credentialid = Integer.parseInt(credentialId);
         Credential credential = credentialMapper.getCredential(credentialid);
+        if (!checkAuthorization(credential.getUserId(), authentication)) {
+            return "redirect:/home?msg=authException";
+        }
         credential.setPassword(encryptionService.decryptValue(credential.getPassword(), credential.getKey()));
         model.addAttribute("credential", credential);
         return "edit-credential";
     }
 
     @PostMapping("/update-credential")
-    public String updateCredential(@ModelAttribute("credential") Credential credential) {
+    public String updateCredential(@ModelAttribute("credential") Credential credential, Authentication authentication) {
         String encryptedPassword = encryptionService.encryptValue(credential.getPassword(), credential.getKey());
         credential.setPassword(encryptedPassword);
+        if (!checkAuthorization(credential.getUserId(), authentication)) {
+            return "redirect:/home?msg=authException";
+        }
         credentialMapper.update(credential);
         return "redirect:/home?msg=uCred";
     }
 
     @PostMapping("/uploadFile")
     public String addFile(@RequestParam("file") MultipartFile fileUpload, Model model, Authentication authentication) throws IOException {
-        User user = userService.getUser(authentication.getName());
-        int userId = user.getUserId();
         if (fileUpload.isEmpty()) {
             model.addAttribute("success", false);
             model.addAttribute("message", "No file selected to upload!");
@@ -186,7 +208,7 @@ public class HomeController {
         String fileExt = com.google.common.io.Files.getFileExtension(fn);
         long fileSize = Files.size(uploadPath);
 
-        fileMapper.addFile(new UploadFile(fn, fileExt, String.valueOf(fileSize), userId, uploadPath.toString()));
+        fileMapper.addFile(new UploadFile(fn, fileExt, String.valueOf(fileSize), userService.getUserId(authentication.getName()), uploadPath.toString()));
 
         model.addAttribute("success", true);
         model.addAttribute("message", "New File added successfully!");
@@ -195,13 +217,16 @@ public class HomeController {
     }
 
     @GetMapping("/uploadFile/delete")
-    public String deleteFile(@RequestParam("fileId") int fileId) {
+    public String deleteFile(@RequestParam("fileId") int fileId, Authentication authentication) {
 
         // delete file on hard disk
 
         UploadFile file = fileMapper.getFile(fileId);
-
         Path filePath = Paths.get(file.getFileLocation());
+
+        if (!checkAuthorization(file.getUserId(), authentication)) {
+            return "redirect:/home?msg=authException";
+        }
 
         try {
             Files.deleteIfExists(filePath);
@@ -217,13 +242,17 @@ public class HomeController {
 
     @GetMapping(value = "/uploadFile/{filename}", produces = MediaType.APPLICATION_OCTET_STREAM_VALUE)
     public void downloadFile(@PathVariable("filename") String filename,
-                             HttpServletResponse response) throws IOException {
+                             HttpServletResponse response, Authentication authentication) throws IOException {
         try {
-            Path uploadDir = Paths.get("./src/main/resources/upload");
-            String filePath = Paths.get(String.valueOf(uploadDir), filename).toString();
-            InputStream is = new FileInputStream(new File(filePath));
+            UploadFile uploadFile = fileMapper.getFileByName(filename);
+
+            if (!checkAuthorization(uploadFile.getUserId(), authentication)) {
+                throw new AuthorizationException("Yetkisiz işlem");
+            }
+            
+            InputStream is = new FileInputStream(new File(uploadFile.getFileLocation()));
             IOUtils.copy(is, response.getOutputStream());
-            String fileExt = com.google.common.io.Files.getFileExtension(filePath);
+            String fileExt = com.google.common.io.Files.getFileExtension(uploadFile.getFileLocation());
             response.setHeader("Content-Disposition", "attachment;filename=download." + fileExt);
             response.flushBuffer();
         } catch (IOException e) {
@@ -232,4 +261,10 @@ public class HomeController {
         }
     }
 
+    private boolean checkAuthorization(Integer userId, Authentication authentication) {
+        if (userId.equals(userService.getUserId(authentication.getName()))) {
+            return true;
+        }
+        return false;
+    }
 }
